:building_construction: Move Search Metadata Routes to Search Module
