+++ src/Search/ExpressRouteAdapter.ts
@@ -1,5 +1,5 @@
 import {OutcomeFilter} from '../Shared/OutcomeFilter';
-import {searchOutcomes} from './SearchInteractor';
+import {fetchAreas, fetchSources, searchOutcomes} from './SearchInteractor';
 import {StandardOutcome} from '@cyber4all/clark-entity';
 import {Router} from 'express';
 import {MongoSearchGateway} from './MongoSearchGateway';
@@ -31,6 +31,24 @@ export function buildRouter() {
             res.send(outcomePayload);
         } catch (e) {
             res.status(500).send(e);
+        }
+    });
+    router.get('/outcomes/sources', async (req, res) => {
+        try {
+            const sources = await fetchSources(this.dataStore);
+            // TODO: Should this be JSON?
+            res.status(200).send(sources);
+        } catch (e) {
+            console.error(e);
+            res.sendStatus(500);
+        }
+    });
+    router.get('/outcomes/areas', async (req, res) => {
+        try {
+            const areas = await fetchAreas(this.dataStore);
+            res.json(areas);
+        } catch (e) {
+            res.status(500).send('Internal Server Error');
         }
     });
     return router;
