+++ src/Search/MongoSearchGateway.ts
@@ -1,11 +1,14 @@
 import {Collection} from 'mongodb';
 import {StandardOutcomeDocument} from '@cyber4all/clark-schema';
 import {MongoConnector} from '../Shared/MongoConnector';
-import {COLLECTIONS} from '../drivers/MongoDriver';
 import {OutcomeFilter} from '../Shared/OutcomeFilter';
 import {StandardOutcome} from '@cyber4all/clark-entity';
+import {OutcomeGateway} from './SearchInteractor';
 
-export class MongoSearchGateway {
+const COLLECTIONS = {
+    STANDARD_OUTCOMES: 'outcomes',
+};
+export class MongoSearchGateway implements OutcomeGateway {
     private standardOutcomes: Collection<StandardOutcomeDocument>;
 
     constructor() {
@@ -35,13 +38,13 @@ export class MongoSearchGateway {
             const skip = page && limit ? (page - 1) * limit : undefined;
             const query: any = {
                 $or: [
-                    { $text: { $search: filter.text } },
-                    { outcome: new RegExp(filter.text, 'ig') },
+                    {$text: {$search: filter.text}},
+                    {outcome: new RegExp(filter.text, 'ig')},
                 ],
             };
             delete filter.text;
             for (const prop of Object.keys(filter)) {
-                query[prop] = { $regex: new RegExp(filter[prop], 'ig') };
+                query[prop] = {$regex: new RegExp(filter[prop], 'ig')};
             }
             let docs = await this.standardOutcomes
                 .find(query);
@@ -58,14 +61,47 @@ export class MongoSearchGateway {
             return {
                 total: total,
                 outcomes: outcomes.map(outcome => new StandardOutcome({
-                        id: outcome._id,
-                        date: `${outcome.date}`,
-                        author: outcome.author,
-                        name: outcome.name,
-                        source: outcome.source,
-                        outcome: outcome.outcome,
-                    })),
+                    id: outcome._id,
+                    date: `${outcome.date}`,
+                    author: outcome.author,
+                    name: outcome.name,
+                    source: outcome.source,
+                    outcome: outcome.outcome,
+                })),
             };
+        } catch (e) {
+            return Promise.reject(e);
+        }
+    }
+
+    /**
+     * Fetches array of distinct sources
+     *
+     * @returns {Promise<string[]>}
+     * @memberof MongoDriver
+     */
+    public async fetchSources(): Promise<string[]> {
+        try {
+            return (<any>(
+                this.standardOutcomes
+            )).distinct('source');
+        } catch (e) {
+            return Promise.reject(e);
+        }
+    }
+
+    /**
+     * Returns all areas of standard outcomes, grouped by source.
+     *
+     * NOTE: "area" is stored as the property "name" in the database.
+     */
+    public async fetchAreas(): Promise<{ _id: string, areas: string[]}[]> {
+        try {
+            // @ts-ignore FIXME: This needs a better solution
+            return this.standardOutcomes
+                .aggregate([
+                    {$group: {_id: '$source', areas: {$addToSet: '$name'}}},
+                ]).toArray();
         } catch (e) {
             return Promise.reject(e);
         }
