+++ src/Search/SearchInteractor.ts
@@ -1,3 +1,7 @@
+import {StandardOutcome} from '@cyber4all/clark-entity';
+import {sanitizeFilter} from '../Shared/SanitizeFilter';
+import {OutcomeFilter} from '../Shared/OutcomeFilter';
+
 /**
  * Searches Outcomes
  *
@@ -8,11 +12,6 @@
  * @param {number} [page]
  * @returns {Promise<{ total: number; outcomes: StandardOutcome[] }>}
  */
-import {DataStore} from '../interfaces/DataStore';
-import {StandardOutcome} from '@cyber4all/clark-entity';
-import {sanitizeFilter} from '../Shared/SanitizeFilter';
-import {OutcomeFilter} from '../Shared/OutcomeFilter';
-
 export async function searchOutcomes(params: {
     dataStore: OutcomeGateway;
     filter: OutcomeFilter;
@@ -31,10 +30,40 @@ export async function searchOutcomes(params: {
     }
 }
 
+/**
+ * Fetches all Standard Outcome sources
+ *
+ * @static
+ * @param {OutcomeGateway} dataStore
+ * @returns {Promise<string[]>}
+ */
+export async function fetchSources(dataStore: OutcomeGateway): Promise<string[]> {
+    try {
+        return dataStore.fetchSources();
+    } catch (e) {
+        return Promise.reject(`Problem finding sources. Error: ${e}.`);
+    }
+}
+
+/**
+ * Fetches all areas of each standard outcome, grouped by source.
+ *
+ * @param dataStore the gateway to the outcome datastore
+ */
+export async function fetchAreas(dataStore: OutcomeGateway): Promise<{ _id: string, areas: string[] }[]> {
+    try {
+        return dataStore.fetchAreas();
+    } catch (e) {
+        return Promise.reject(`Problem finding sources. Error: ${e}.`);
+    }
+}
+
 export interface OutcomeGateway {
     searchOutcomes(
         filter: OutcomeFilter,
         limit?: number,
         page?: number,
     ): Promise<{ total: number; outcomes: StandardOutcome[] }>;
+    fetchSources(): Promise<string[]>;
+    fetchAreas(): Promise<{ _id: string, areas: string[]}[]>;
 }
