+++ src/Suggestion/SuggestMode.ts
@@ -1,0 +1,1 @@
+export type suggestMode = 'text' | 'regex';
