+++ src/app.ts
@@ -1,6 +1,6 @@
-import { ExpressDriver } from './drivers/drivers';
 import * as dotenv from 'dotenv';
 import { MongoConnector } from './Shared/MongoConnector';
+import {ExpressDriver} from './express/ExpressDriver';
 dotenv.config();
 // ----------------------------------------------------------------------------------
 // Initializations
