--- src/drivers/drivers.ts
@@ -1,4 +1,0 @@
-import { ExpressDriver } from './express/ExpressDriver';
-import { ExpressRouteDriver } from './express/ExpressRouteDriver';
-import { MongoDriver } from './MongoDriver';
-export { ExpressDriver, ExpressRouteDriver, MongoDriver };
