--- src/interfaces/DataStore.ts
@@ -1,6 +1,0 @@
-export interface DataStore {
-  fetchSources(): Promise<string[]>;
-  fetchAreas(): Promise<{ _id: string, areas: string[]}[]>;
-}
-
-export type suggestMode = 'text' | 'regex';
