--- src/types/standard-outcome-document.ts
@@ -1,9 +1,0 @@
-export interface StandardOutcomeDocument {
-  _id: string;
-  author: string;
-  name: string;
-  date: string;
-  outcome: string;
-  source: string;
-  tag: string;
-}
